[IMP] runbot: add a _json_data field on build
When two steps in the same build needs to exchange informations, some
hacks have to be used. E.g. using the extra_params fields to store comma
separated values.

With this commit, a _json_data field is added alongside with a json_data
property that automatically transform the data into json.