[FIX] runbot: correctly import os when using cli
When container.py is used from command line, it crashes because it
cannot import os from the "common" file.

The container started by the CLI now use the new docker_state function
to check if the container is running or has finished.

Also, this commit adds test-enable params to the container.py script which is
symmetric to the same Odoo CLI parameter.