[FIX] runbot: handle git archive failure
When a failure occurs during the export of the commit, if the failure is
at the git command level the error message is lost in the system logs.

At the user level, the only feedback is that no server was found in the
sources, which is a bit obscure.

With this commit, the return code of the git process is verified and an
Exception is raised in case of failure.