[FIX] runbot: sanitize build dest also
In a previous commit 9684c2d, the docker name was sanitized to prevent
crash when the docker name contains unallowed characters.

Unfortunately, some characters are able to create problems elsewhere.
For example, when a database is downloaded, if a `#` is present in the
branch name, the zip filename will contains this character too, leading
to 404 when trying to download the zip.

To avoid further problems, this fix use the function from the previous
commit to sanitize the build dest, as many other names are computed from
the dest.