[FIX] *: handle rate limit in pylint tests
With a concurrency of 3 or more, it ends up being pretty easy to hit
github's rate limit (5000 requests/h), at which point the run hits a
cascading failure where every test from thereon blows up to the rate
limiting.

Add a handling for that case in some of the early github-querying
fixtures, so they can wait for the ratelimit delay to be restored.

This increases the need for a proper fake github thingie I could run
on a per-test basis.