[FIX] runbot bundle missing base_id
When trying to set up my runbot instance, I ran into an issue where the base_id of a bundle
was sometimes empty. I'm not sure exactly what causes this, possibly it happens when
bootstrapping a runbot from scratch, or there is an issue with the triggers of the computed
field. The net result was that my PR branches had no base_id in their bundles and would not build.

This PR fixes the issue by explicitly calling the _compute_base_id method when the bundle has no version
or no base id.